HTML5 - offscreen frramebuffer mode (multi-threading) - switching to fullscreen mode crash fix
#jira UEMOB-255  HTML5 Multithreaded rendering with RHIT
#rn
#rb none

[CL 6623719 by Nick Shin in Dev-Mobile branch]

+++ Engine/Extras/ThirdPartyNotUE/emsdk/emscripten/1.38.31/src/library_webgl.js
@@ -1624,6 +1624,9 @@ var LibraryGL = {
       // From spec: "glDeleteBuffers silently ignores 0's and names that do not
       // correspond to existing buffer objects."
       if (!buffer) continue;
+// EPIC EDIT start -- nick.shin 2019-02-06 -- UEMOB-255
+      if (!GLctx) continue;		// this check is needed when using OFFSCREEN_FRAMEBUFFER
+// EPIC EDIT end -- nick.shin 2019-02-06
 
       GLctx.deleteBuffer(buffer);
       buffer.name = 0;
